Created a function for create object
