+++ packages/custom-props-layer/src/components/text/Text.tsx
@@ -3,6 +3,7 @@ import { useMemo, useCallback } from "react";
 import { Label } from "../commons/Label";
 import { PropertyContainer } from "../commons/PropertyContainer";
 import { TextInput } from "../commons/TextInput";
+import { createObject } from "../../utility/Utility";
 
 export const Text: React.FC<ComponentProps> = (props) => {
   const selector = useMemo(() => {
@@ -18,17 +19,6 @@ export const Text: React.FC<ComponentProps> = (props) => {
 
   const callPatchCb = useCallback(
     (e: React.ChangeEvent<HTMLInputElement>) => {
-      const createObject = (fields: string[], value: string) => {
-        const reducer: any = (
-          acc: string,
-          item: string,
-          index: number,
-          arr: string[]
-        ) => ({
-          [item]: index + 1 < arr.length ? acc : value,
-        });
-        return fields.reduceRight(reducer, {});
-      };
       props.patchCb({
         property: {
           custom: createObject(selector, e.target.value),
