+++ packages/custom-props-layer/src/utility/Utility.ts
@@ -1,0 +1,11 @@
+export const createObject = (fields: string[], value: string | number) => {
+  const reducer: any = (
+    acc: string,
+    item: string,
+    index: number,
+    arr: string[]
+  ) => ({
+    [item]: index + 1 < arr.length ? acc : value,
+  });
+  return fields.reduceRight(reducer, {});
+};
